defense_service/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
numpy
pymongo
requests
python-multipart
python-dotenv
openai>=1.0.0
